Email Validation/requirements.txt
flask
dnspython
tqdm
